Add rating feature, refactor ShoppingCart classes
This commit primarily adds the rating feature by introducing new classes RatingController and RatingService.  It also includes another important change where "Book" was replaced with "Books" in the ShoppingCart classes. Furthermore, the terms Rating and Book have been edited to follow naming conventions in the Rating.java file. The RatingRepo.java file was extended to include more features in line with the introduction of the rating feature.